xplat: use Single library for static builds of ChakraCore.
This commit takes the changes introduced in the previous commit and makes it
to where all internal cross dependencies of ChakraCore are satisfied with a
single static library called libChakraCoreStatic.a.

NOTE: you still need to make sure to link against all required system
libraries to ensure all symbol requirements are met.